+++ src/main/java/com/webank/webase/node/mgr/contract/abi/entity/ReqImportAbi.java
@@ -31,7 +31,7 @@ import java.util.List;
 public class ReqImportAbi {
 	private Integer abiId;
 	@NotBlank
-    private String account;
+	private String account;
 	@NotNull
 	private String groupId;
 	@NotBlank
@@ -40,5 +40,9 @@ public class ReqImportAbi {
 	private String contractAddress;
 	@NotNull
 	private List<Object> contractAbi;
-//	private String deployAddress;
+
+	/**
+	 * 0-solidity, 1-liquid
+	 */
+	private Boolean isWasm = false;
 }