+++ src/main/java/com/webank/webase/node/mgr/front/entity/TransactionParam.java
@@ -40,4 +40,5 @@ public class TransactionParam {
     private boolean useCns = false;
     private String cnsName;
     private String version;
+    private Boolean isWasm = false;
 }