+++ src/main/java/com/webank/webase/node/mgr/monitor/MonitorService.java
@@ -56,8 +56,10 @@ import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang3.StringUtils;
+import org.fisco.bcos.sdk.codec.datatypes.Address;
 import org.fisco.bcos.sdk.codec.wrapper.ABIDefinition;
 import org.fisco.bcos.sdk.crypto.CryptoSuite;
+import org.fisco.bcos.sdk.model.TransactionReceipt;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -95,6 +97,8 @@ public class MonitorService {
     @Lazy
     private AbiService abiService;
 
+    private final static List<String> LIQUID_PRECOMPILED_ADDRESS_ARRAY = Arrays.asList("/sys/auth", "/sys/bfs", "/sys/cns", "/sys/consensus",
+        "/sys/crypto_tools", "/sys/kv_storage", "/sys/parallel_config", "/sys/status", "/sys/table_storage");
 
     /**
      * monitor every group.
@@ -387,15 +391,44 @@ public class MonitorService {
         String contractAddress, contractName, interfaceName = "", contractBin;
         int transType = TransType.DEPLOY.getValue();
         int transUnusualType = TransUnusualType.NORMAL.getValue();
+        // liquid时to在部署和调用时均不为空，需要获取回执判断contractAddress
+        if (transTo.startsWith("/")) {
+            TransactionReceipt receipt = frontInterface.getTransReceipt(groupId, transHash);
+            contractAddress = receipt.getContractAddress();
+            // 部署失败的时候，或者调用交易的时候，address为空
+            if (StringUtils.isBlank(contractAddress)) {
+                if (!receipt.isStatusOK()) {
+                    // 部署失败
+                    return new ContractMonitorResult("0x", "0x", TransType.DEPLOY.getValue(),
+                        MonitorUserType.NORMAL.getValue());
+                } // else 发交易，执行下文根据transTo判断
+                // continue
+            } else {
+                //部署成功，手动设置transTo为空，适配solidity中的to为空，contractAddress非空的特点
+                transTo = "";
+            }
+        }
         // deploy contract tx
         if (StringUtils.isBlank(transTo) || "0x".equalsIgnoreCase(transTo)) {
             contractAddress = frontInterface.getAddressByHash(groupId, transHash);
             // if contract deploy error, contract address is null and transTo is null
-            if (StringUtils.isBlank(contractAddress)) {
-                log.warn("transTo is empty, and contract address is empty for deploy error");
-                return new ContractMonitorResult("0x", "0x", MonitorUserType.NORMAL.getValue());
+            if (StringUtils.isBlank(contractAddress) || Address.DEFAULT.getValue().equalsIgnoreCase(contractAddress) ) {
+                log.warn("transTo is empty, and contract address is empty because deploy error");
+                return new ContractMonitorResult("0x", "0x", TransType.DEPLOY.getValue(),
+                    MonitorUserType.NORMAL.getValue());
             }
-            contractBin = frontInterface.getCodeFromFront(groupId, contractAddress, blockNumber);
+            if (contractAddress.startsWith("0x0000000000000000000000000000000000")
+                || isPrecompiledLiquidAddress(contractAddress)) {
+                log.info("contractAddress is precompiled contract, skip");
+                return new ContractMonitorResult(contractAddress, contractAddress, TransType.DEPLOY.getValue(),
+                    MonitorUserType.NORMAL.getValue());
+            }
+            contractBin = frontInterface.getCodeV2FromFront(groupId, contractAddress, blockNumber);
+            if (StringUtils.isBlank(contractBin)) {
+                log.warn("contractAddress:[{}] not exist on chain, required audit", contractAddress);
+                return new ContractMonitorResult(contractAddress, contractAddress, TransType.CALL.getValue(),
+                    MonitorUserType.ABNORMAL.getValue());
+            }
             contractBin = removeBinFirstAndLast(contractBin);
 
             List<TbContract> contractRow = contractService.queryContractByBin(groupId, contractBin);
@@ -417,7 +450,19 @@ public class MonitorService {
             transType = TransType.CALL.getValue();
             String methodId = transInput.substring(0, 10);
             contractAddress = transTo;
-            contractBin = frontInterface.getCodeFromFront(groupId, contractAddress, blockNumber);
+            if (contractAddress.startsWith("0x0000000000000000000000000000000000")
+                || isPrecompiledLiquidAddress(contractAddress) ) {
+                log.info("contractAddress is precompiled contract, skip");
+                return new ContractMonitorResult(contractAddress, contractAddress, TransType.CALL.getValue(),
+                    MonitorUserType.NORMAL.getValue());
+            }
+            contractBin = frontInterface
+                .getCodeV2FromFront(groupId, contractAddress, blockNumber);
+            if (StringUtils.isBlank(contractBin)) {
+                log.warn("contractAddress:[{}] not exist on chain, required audit", contractAddress);
+                return new ContractMonitorResult(contractAddress, contractAddress, TransType.CALL.getValue(),
+                    MonitorUserType.ABNORMAL.getValue());
+            }
             contractBin = removeBinFirstAndLast(contractBin);
 
             List<TbContract> contractRow = contractService.queryContractByBin(groupId, contractBin);
@@ -531,6 +576,9 @@ public class MonitorService {
      * get contractName from contractBin.
      */
     private String getNameFromContractBin(String groupId, String contractBin) {
+        if (StringUtils.isBlank(contractBin)) {
+            return null;
+        }
         List<TbContract> contractList = contractService.queryContractByBin(groupId, contractBin);
         if (contractList != null && contractList.size() > 0) {
             return contractList.get(0).getContractName();
@@ -547,5 +595,12 @@ public class MonitorService {
             contractName = contractBin.substring(contractBin.length() - 10);
         }
         return contractName;
+    }
+
+    private static boolean isPrecompiledLiquidAddress(String liquidAddress) {
+        if (StringUtils.isBlank(liquidAddress)) {
+            return false;
+        }
+        return LIQUID_PRECOMPILED_ADDRESS_ARRAY.stream().anyMatch(liquidAddress::equalsIgnoreCase);
     }
 }