+++ src/main/java/com/webank/webase/node/mgr/precntauth/authmanager/everyone/entity/ReqProposalInfo.java
@@ -1,0 +1,11 @@
+package com.webank.webase.node.mgr.precntauth.authmanager.everyone.entity;
+
+import lombok.Data;
+
+@Data
+public class ReqProposalInfo {
+
+  private String groupId;
+  private String proposalId;
+
+}