+++ src/main/java/com/webank/webase/node/mgr/precntauth/precompiled/consensus/entity/ReqNodeListInfo.java
@@ -1,0 +1,13 @@
+package com.webank.webase.node.mgr.precntauth.precompiled.consensus.entity;
+
+import lombok.Data;
+
+@Data
+public class ReqNodeListInfo {
+
+  private String groupId;
+  private Integer pageSize;
+  private Integer pageNumber;
+
+
+}