+++ src/main/java/com/webank/webase/node/mgr/tools/PrecompiledTools.java
@@ -17,7 +17,7 @@ package com.webank.webase.node.mgr.tools;
 import com.webank.webase.node.mgr.base.code.ConstantCode;
 import com.webank.webase.node.mgr.base.code.RetCode;
 import com.webank.webase.node.mgr.base.entity.BaseResponse;
-import com.webank.webase.node.mgr.precompiled.entity.PrecompiledResult;
+import com.webank.webase.node.mgr.precntauth.authmanager.base.PrecompiledResult;
 
 /**
  * check precompiled api return code