+++ Changelog.md
@@ -1,4 +1,18 @@
 
+### lab-rc2(2022-05-07)
+
+**Add**
+- 支持Liquid合约功能
+- 支持BFS、CNS、权限管理、Event查看与事件订阅等功能
+
+**兼容性**
+- 支持FISCO-BCOS v3.0.0-rc2 及以上版本
+- WeBASE-Web lab-rc2
+- WeBASE-Front lab-rc2
+- WeBASE-Sign lab-rc1
+
+详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/lab/)。
+
 ### lab-rc1(2021-12-09)
 
 **Add**
@@ -10,6 +24,7 @@
 - WeBASE-Web lab-rc1
 - WeBASE-Front lab-rc1
 - WeBASE-Sign lab-rc1
+
 详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/lab/)。